Update the code and remove conditional which is not the best idea.
This way we still exercuse and cover that code path as part of the
tests.

We now disable logging of the very noise logging of the raw payload
inside the tests by utilizing a separate logger when logging a payload.

This may be a good idea in general since it also allows us to
potentially route those logs to a different file or similar.